requirements.txt
pandas
playwright
webdriver-manager
bs4
redis
python-dotenv